feat: wire database migrate endpoint + gate update center scheduler
Fix 1 (P2): POST /api/settings/database/migrate no longer returns 501.
The migration logic is extracted from cmd/migrate/main.go into a callable
store.RunMigration(opts) (store/migrate_runner.go), parameterising the log
sink via opts.LogWriter so the same code path serves both the CLI and the
admin handler. The handler validates the request, resolves the live runtime
DB as the source, refuses an in-place overwrite, then queues a
"database_migration" background task (dedupe key) that streams RunMigration
progress into /api/tasks/{id} via the new AppendBackgroundTaskLog helper,
returning 202 Accepted + the task ID. cmd/migrate is now a thin flag wrapper.

Fix 2 (P3): the update-center scheduler (a log-only no-op) is gated behind
METAPI_ENABLE_UPDATE_CENTER (default false) in app/services.go so it stops
taking a 15-min lease with no product value; it logs a clear disabled message
and is still registered when the flag is set.

Tests: migrate handler (202 + task ID + terminal state + same-target guard +
empty-string guard), store.RunMigration SQLite round-trip + dry-run + moved
isPostgresURL/builders-drift tests, update-center gating (off by default, on
when flag set), config flag defaults-off + parse.
