+++ app/services.go
@@ -99,7 +99,14 @@ func buildSchedulers(cfg *config.Config) (
 	newRegistry.Register(scheduler.NewSub2APIRefreshScheduler(cfg))
 
 	// ---- Scheduler 10: Update Center ----
-	newRegistry.Register(scheduler.NewUpdateCenterScheduler(cfg))
+	// Gated behind METAPI_ENABLE_UPDATE_CENTER (default false). The scheduler
+	// is a log-only no-op today (no remote registry / version discovery), so it
+	// is registered only when an operator explicitly opts in.
+	if cfg.UpdateCenterEnabled {
+		newRegistry.Register(scheduler.NewUpdateCenterScheduler(cfg))
+	} else {
+		slog.Info("update center scheduler disabled (METAPI_ENABLE_UPDATE_CENTER not set)")
+	}
 
 	// ---- Scheduler 12: Admin Snapshot ----
 	newRegistry.Register(scheduler.NewAdminSnapshotScheduler(cfg, usageAgg))
